requirements.txt
lxml==4.6.3
Pillow==8.2.0
python-docx==0.8.11
